feat: make opencv test use cone scan pipeline
